requirements.txt
# Linux
# sudo apt-get install python-twisted

pillow
psycopg2
service_identity
django==1.5.12
django-htmlmin==0.7.0
django_compressor==1.4
django-grappelli==2.6.3
django-getenv==1.3.1
django-debug-toolbar==1.3.0
Scrapy==0.24.5
django-allauth==0.12.0
django-bootstrap3==2.5.6
django-widget-tweaks==1.4.1
pytz
twistar
